Relocate server constructor to folder. Write Gulp task to rebuild server.
